Phoenix LiveView (LiveComponent) Fixes #214
