+++ day16-17/blog_app/lib/blog_app_web/router.ex
@@ -73,8 +73,8 @@ defmodule BlogAppWeb.Router do
 
     live_session :require_authenticated_account,
       on_mount: [{BlogAppWeb.AccountAuth, :ensure_authenticated}] do
-      live "/accounts/settings", AccountSettingsLive, :edit
-      live "/accounts/settings/confirm_email/:token", AccountSettingsLive, :confirm_email
+      live "/accounts/settings", AccountPageLive, :edit
+      live "/accounts/settings/confirm_email/:token", AccountPageLive, :confirm_email
 
       live "/articles/new", ArticleLive.Form, :new
       live "/articles/:article_id/edit", ArticleLive.Form, :edit
